feat(ui): redesign settings as a full-page surface (impl 0025)
Replace the 680x560 settings modal with a full-window /settings/:pane?
route: resizable sidebar (shared width store with the app sidebar),
grouped nav with label search, card-styled panes extracted to
src/components/settings/. Adds the read-only Keyboard shortcuts pane
over a static keymap registry (feature 257 v1), merges Updates into
About behind a five-state update button ladder with auto-check on
mount, and replaces UpdateToast with an Arc-style UpdatePromptCard
above the sidebar Settings row. Entry points: sidebar row, command
palette entry, and Cmd+, (bridged through xterm focus). Update checks
now only start from resting updater states so overlapping auto-checks
can't clobber an in-flight download.
