--- src/components/SettingsModal.tsx
@@ -1,1669 +1,0 @@
-// Settings modal — Pencil shells `hnxWB` (General default), `Wx8dI`
-// (Updates), `Ohaky` (About). Sidebar nav on the left + per-pane
-// content on the right.
-//
-// All settings persist to localStorage for now: there's no backend
-// settings store yet, but the surfaces are in place so individual
-// settings can land without UI churn. "Default working directory"
-// is read by StartMissionModal, CreateRunnerModal, and the direct-
-// chat spawn sites via the helpers in `src/lib/settings.ts`;
-// "Default crew" still has no consumer (follow-up).
-//
-// Entry point: AppShell mounts a button (`Settings` link in the
-// sidebar) that toggles `open`.
-
-import { useEffect, useRef, useState } from "react";
-import {
-  BookText,
-  Check,
-  Copy,
-  Download,
-  ExternalLink,
-  FileText,
-  FolderOpen,
-  Info,
-  Loader2,
-  MessageSquare,
-  Minus,
-  Monitor,
-  Moon,
-  Plus,
-  Plug,
-  RefreshCw,
-  RotateCcw,
-  Scale,
-  Settings as SettingsIcon,
-  ShieldAlert,
-  Sun,
-  Terminal,
-  X,
-} from "lucide-react";
-
-import { invoke } from "@tauri-apps/api/core";
-import { openUrl } from "@tauri-apps/plugin-opener";
-import { getVersion } from "@tauri-apps/api/app";
-
-import { api } from "../lib/api";
-import { applyAppZoom } from "../lib/appZoom";
-import appIcon from "../assets/app-icon.png";
-import {
-  APP_FONT_FAMILY_OPTIONS,
-  APP_THEME_OPTIONS,
-  DARK_VARIANT_ACCENTS,
-  DARK_VARIANT_LABELS,
-  DARK_VARIANT_OPTIONS,
-  LIGHT_VARIANT_ACCENTS,
-  LIGHT_VARIANT_LABELS,
-  LIGHT_VARIANT_OPTIONS,
-  notifySameWindowStorage,
-  readAppFontFamily,
-  readAppTheme,
-  readAppZoom,
-  readDarkVariant,
-  readDefaultChatRuntime,
-  readDefaultWorkingDir,
-  readLightVariant,
-  readStoredBool,
-  readTerminalCursorStyle,
-  readTerminalFontFamily,
-  readTerminalFontSize,
-  readTerminalScrollback,
-  readTerminalTheme,
-  resolveAppSurface,
-  STORAGE_APP_DARK_VARIANT,
-  STORAGE_APP_FONT_FAMILY,
-  STORAGE_APP_LIGHT_VARIANT,
-  STORAGE_APP_THEME,
-  STORAGE_APP_ZOOM,
-  STORAGE_AUTO_INSTALL_UPDATES,
-  STORAGE_TERMINAL_CURSOR_STYLE,
-  STORAGE_TERMINAL_FONT_FAMILY,
-  STORAGE_TERMINAL_FONT_SIZE,
-  STORAGE_TERMINAL_SCROLLBACK,
-  STORAGE_TERMINAL_THEME,
-  TERMINAL_CURSOR_STYLE_OPTIONS,
-  TERMINAL_FONT_FAMILY_OPTIONS,
-  TERMINAL_FONT_SIZE_MAX,
-  TERMINAL_FONT_SIZE_MIN,
-  TERMINAL_SCROLLBACK_OPTIONS,
-  TERMINAL_THEME_ACCENTS,
-  TERMINAL_THEME_LABELS,
-  TERMINAL_THEME_OPTIONS,
-  type AppFontFamily,
-  type AppTheme,
-  type DarkVariant,
-  type LightVariant,
-  type TerminalCursorStyle,
-  type TerminalFontFamily,
-  type TerminalTheme,
-  writeAppFontFamily,
-  writeAppTheme,
-  writeDarkVariant,
-  writeDefaultChatRuntime,
-  writeDefaultWorkingDir,
-  writeLightVariant,
-  writeStoredBool,
-  writeTerminalCursorStyle,
-  writeTerminalFontFamily,
-  writeTerminalFontSize,
-  writeTerminalScrollback,
-  writeTerminalTheme,
-  ZOOM_STEPS,
-} from "../lib/settings";
-import { useUpdate } from "../contexts/UpdateContext";
-import { StyledSelect } from "./ui/StyledSelect";
-import { WorkingDirField } from "./ui/WorkingDirField";
-
-interface SettingsModalProps {
-  open: boolean;
-  onClose: () => void;
-}
-
-type Pane =
-  | "general"
-  | "chat"
-  | "appearance"
-  | "terminal"
-  | "mcp"
-  | "diagnostics"
-  | "updates"
-  | "about";
-
-const PANES: { key: Pane; label: string; subtitle: string; icon: typeof SettingsIcon }[] = [
-  {
-    key: "general",
-    label: "General",
-    subtitle: "Startup & defaults",
-    icon: SettingsIcon,
-  },
-  {
-    key: "chat",
-    label: "Chat",
-    subtitle: "Direct chat defaults",
-    icon: MessageSquare,
-  },
-  {
-    key: "appearance",
-    label: "Appearance",
-    subtitle: "Theme & sidebar mark",
-    icon: Sun,
-  },
-  {
-    key: "terminal",
-    label: "Terminal",
-    subtitle: "xterm appearance",
-    icon: Terminal,
-  },
-  {
-    key: "mcp",
-    label: "MCP",
-    subtitle: "Client bindings",
-    icon: Plug,
-  },
-  {
-    key: "diagnostics",
-    label: "Diagnostics",
-    subtitle: "Logs & troubleshooting",
-    icon: FileText,
-  },
-  {
-    key: "updates",
-    label: "Updates",
-    subtitle: "Channel & auto-update",
-    icon: Download,
-  },
-  { key: "about", label: "About", subtitle: "Version & links", icon: Info },
-];
-
-type McpClientId = "claude_code" | "codex";
-type McpCopyKey = McpClientId | "binding_dir";
-
-interface McpClientStatus {
-  registered: boolean;
-  matches_current: boolean;
-  command: string | null;
-  args: string[];
-  config_path: string;
-  error: string | null;
-}
-
-interface McpIntegrationStatus {
-  environment: string;
-  binary_path: string;
-  socket_path: string;
-  claude_code: McpClientStatus;
-  codex: McpClientStatus;
-}
-
-interface McpConfigSnippet {
-  claude_code: string;
-  codex: string;
-}
-
-export function SettingsModal({ open, onClose }: SettingsModalProps) {
-  const [pane, setPane] = useState<Pane>("general");
-  const cardRef = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    if (!open) return;
-    const onKey = (e: KeyboardEvent) => {
-      if (e.key === "Escape") onClose();
-    };
-    const onMouseDown = (e: MouseEvent) => {
-      if (!cardRef.current) return;
-      if (!cardRef.current.contains(e.target as Node)) onClose();
-    };
-    document.addEventListener("keydown", onKey);
-    document.addEventListener("mousedown", onMouseDown);
-    return () => {
-      document.removeEventListener("keydown", onKey);
-      document.removeEventListener("mousedown", onMouseDown);
-    };
-  }, [open, onClose]);
-
-  if (!open) return null;
-
-  return (
-    <div className="fixed inset-0 z-50 flex items-center justify-center bg-black/55 p-4">
-      <div
-        ref={cardRef}
-        className="flex h-[85vh] max-h-[640px] w-[90vw] max-w-[760px] overflow-hidden rounded-xl border border-line bg-panel shadow-[0_14px_40px_rgba(0,0,0,0.6)]"
-      >
-        {/* Sidebar — Codex-style: secondary surface (panel) sitting
-            alongside the primary surface (bg) used for the content
-            pane. Active item is a neutral darker pill, no accent
-            tint — keeps the sidebar feeling like a flat list rather
-            than highlighting the current section in brand color. */}
-        <aside className="flex w-[200px] shrink-0 flex-col gap-1 border-r border-line bg-sidebar px-3 py-4">
-          <div className="px-2 pb-2 text-[14px] font-bold text-fg">
-            Settings
-          </div>
-          {PANES.map((p) => {
-            const active = pane === p.key;
-            const Icon = p.icon;
-            return (
-              <button
-                key={p.key}
-                type="button"
-                onClick={() => setPane(p.key)}
-                className={`flex cursor-pointer items-center gap-2.5 rounded-lg px-2.5 py-2 text-left transition-colors ${
-                  active
-                    ? "bg-line text-fg"
-                    : "text-fg hover:bg-line/50"
-                }`}
-              >
-                <Icon
-                  aria-hidden
-                  className={`h-4 w-4 ${active ? "text-fg" : "text-fg-2"}`}
-                />
-                <div className="flex min-w-0 flex-col gap-px">
-                  <span className="truncate text-[12px] font-medium text-fg">
-                    {p.label}
-                  </span>
-                  <span
-                    className={`truncate text-[10px] ${
-                      active ? "text-fg-2" : "text-fg-3"
-                    }`}
-                  >
-                    {p.subtitle}
-                  </span>
-                </div>
-              </button>
-            );
-          })}
-        </aside>
-
-        {/* Content — primary surface (bg) so the sidebar's secondary
-            surface (panel) reads as the supporting nav rail. In light
-            mode this puts a white pane next to a gray rail; in dark
-            mode the content is the deeper of the two surfaces. */}
-        <div className="relative flex flex-1 flex-col gap-[18px] overflow-y-auto bg-bg px-6 py-5">
-          <button
-            type="button"
-            onClick={onClose}
-            aria-label="Close settings"
-            title="Close"
-            className="absolute right-3 top-3 flex h-7 w-7 cursor-pointer items-center justify-center rounded-md text-fg-3 hover:bg-raised hover:text-fg"
-          >
-            <X aria-hidden className="h-4 w-4" />
-          </button>
-          {pane === "general" ? <GeneralPane /> : null}
-          {pane === "chat" ? <ChatPane /> : null}
-          {pane === "appearance" ? <AppearancePane /> : null}
-          {pane === "terminal" ? <TerminalPane /> : null}
-          {pane === "mcp" ? <McpPane /> : null}
-          {pane === "diagnostics" ? <DiagnosticsPane /> : null}
-          {pane === "updates" ? <UpdatesPane /> : null}
-          {pane === "about" ? <AboutPane /> : null}
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function PaneHeader({ title, subtitle }: { title: string; subtitle: string }) {
-  return (
-    <div className="flex flex-col gap-1 pr-10">
-      <h2 className="text-[18px] font-semibold text-fg">{title}</h2>
-      <p className="text-[13px] text-fg-2">{subtitle}</p>
-      <div className="mt-3 h-px w-full bg-line" />
-    </div>
-  );
-}
-
-// localStorage-backed boolean. Stubbed here because there's no
-// backend settings store yet — we want the toggle to feel real and
-// persist across reloads, even if no other surface reads the value
-// today.
-function useStoredBool(key: string, initial: boolean): [boolean, (v: boolean) => void] {
-  // Thin wrapper around the shared `lib/settings` helpers so the
-  // modal and any non-React reader (e.g. UpdateContext) can't drift
-  // on encoding. Both sides go through `readStoredBool` /
-  // `writeStoredBool`.
-  const [value, setValue] = useState<boolean>(() => readStoredBool(key, initial));
-  const set = (v: boolean) => {
-    setValue(v);
-    writeStoredBool(key, v);
-  };
-  return [value, set];
-}
-
-function GeneralPane() {
-  // Default crew selector. Persisted to localStorage today (no
-  // backend settings store yet); the StartMissionModal can read the
-  // same key to pre-fill its crew picker once that wiring lands.
-  const [crews, setCrews] = useState<{ id: string; name: string }[]>([]);
-  const [defaultCrewId, setDefaultCrewIdState] = useState<string>(() => {
-    try {
-      return localStorage.getItem("settings.defaultCrewId") ?? "";
-    } catch {
-      return "";
-    }
-  });
-  // Default working directory. Picked via Tauri's dialog plugin
-  // (open({ directory: true })) so the value is always an absolute
-  // path the OS confirmed exists.
-  const [defaultWorkingDir, setDefaultWorkingDirState] = useState<string>(
-    () => readDefaultWorkingDir(),
-  );
-  useEffect(() => {
-    let cancelled = false;
-    void api.crew
-      .list()
-      .then((rows) => {
-        if (cancelled) return;
-        setCrews(rows.map((c) => ({ id: c.id, name: c.name })));
-      })
-      .catch(() => {
-        // best-effort — leave the dropdown empty if the list query
-        // fails; the user can retry by reopening Settings.
-      });
-    return () => {
-      cancelled = true;
-    };
-  }, []);
-  const setDefaultCrewId = (id: string) => {
-    setDefaultCrewIdState(id);
-    try {
-      if (id) localStorage.setItem("settings.defaultCrewId", id);
-      else localStorage.removeItem("settings.defaultCrewId");
-    } catch {
-      // best-effort
-    }
-  };
-  const setDefaultWorkingDir = (path: string) => {
-    setDefaultWorkingDirState(path);
-    writeDefaultWorkingDir(path);
-  };
-  // App zoom — snap-to-step value driven by `ZOOM_STEPS`. Persist + apply
-  // immediately so the user feels the change while picking. The boot-time
-  // apply in `App.tsx` is what makes it survive restarts. Goes through
-  // the shared `applyAppZoom` so the stepper and the global Cmd+/- path
-  // can't drift.
-  const [appZoom, setAppZoomState] = useState<number>(() => readAppZoom());
-  const setAppZoom = (next: number) => {
-    setAppZoomState(next);
-    applyAppZoom(next);
-  };
-  // Keep the visible % in sync when zoom changes from outside the modal
-  // (Cmd+/-/0 shortcut). `applyAppZoom` synthesizes a storage event after
-  // each write so we get a single notification path.
-  useEffect(() => {
-    const onStorage = (e: StorageEvent) => {
-      if (e.key !== STORAGE_APP_ZOOM) return;
-      setAppZoomState(readAppZoom());
-    };
-    window.addEventListener("storage", onStorage);
-    return () => window.removeEventListener("storage", onStorage);
-  }, []);
-  return (
-    <>
-      <PaneHeader title="General" subtitle="Defaults and startup behavior." />
-      <Row
-        label="Default crew"
-        sub="Pre-selected when starting a new mission."
-      >
-        <StyledSelect
-          value={defaultCrewId}
-          options={[
-            { value: "", label: "No default" },
-            ...crews.map((c) => ({ value: c.id, label: c.name })),
-          ]}
-          onChange={setDefaultCrewId}
-        />
-      </Row>
-      <Row
-        label="Default working directory"
-        sub="Cwd new chats inherit unless overridden."
-      >
-        <WorkingDirField
-          singleLine
-          className="w-[280px]"
-          value={defaultWorkingDir}
-          onChange={setDefaultWorkingDir}
-        />
-      </Row>
-      <Row
-        label="App zoom"
-        sub="Whole-app scale. Doesn't apply to the runner terminal canvas — see Terminal pane."
-      >
-        <ZoomStepper value={appZoom} onChange={setAppZoom} />
-      </Row>
-    </>
-  );
-}
-
-function ChatPane() {
-  const [runtimes, setRuntimes] = useState<
-    { name: string; display_name: string; command: string }[]
-  >([]);
-  const [defaultRuntime, setDefaultRuntimeState] = useState<string>(() =>
-    readDefaultChatRuntime(),
-  );
-
-  useEffect(() => {
-    let cancelled = false;
-    void api.runtime
-      .list()
-      .then((rows) => {
-        if (cancelled) return;
-        setRuntimes(rows);
-        const stored = readDefaultChatRuntime();
-        if (stored && !rows.some((runtime) => runtime.name === stored)) {
-          setDefaultRuntimeState("");
-          writeDefaultChatRuntime("");
-        }
-      })
-      .catch(() => {
-        if (!cancelled) setRuntimes([]);
-      });
-    return () => {
-      cancelled = true;
-    };
-  }, []);
-
-  const setDefaultRuntime = (runtime: string) => {
-    setDefaultRuntimeState(runtime);
-    writeDefaultChatRuntime(runtime);
-  };
-
-  return (
-    <>
-      <PaneHeader title="Chat" subtitle="Defaults for direct chats." />
-      <Row
-        label="Default runtime"
-        sub="Pre-selected in Start Chat's Direct mode."
-      >
-        <StyledSelect
-          value={defaultRuntime}
-          options={[
-            { value: "", label: "First available" },
-            ...runtimes.map((runtime) => ({
-              value: runtime.name,
-              label: runtime.display_name,
-            })),
-          ]}
-          onChange={setDefaultRuntime}
-        />
-      </Row>
-    </>
-  );
-}
-
-function AppearancePane() {
-  // Theme = user *intent* (auto/light/dark). lightVariant + darkVariant
-  // are which palette to use on each side; v1 ships Codex Light + Carbon
-  // & Plasma. Writes go through `writeAppTheme` / `writeLightVariant`
-  // then fire a same-window storage event so `applyAppTheme()` in
-  // `main.tsx` flips `<html data-theme>` without a reload.
-  const [theme, setThemeState] = useState<AppTheme>(() => readAppTheme());
-  const [lightVariant, setLightVariantState] = useState<LightVariant>(() =>
-    readLightVariant(),
-  );
-  const [darkVariant, setDarkVariantState] = useState<DarkVariant>(() =>
-    readDarkVariant(),
-  );
-  const [fontFamily, setFontFamilyState] = useState<AppFontFamily>(() =>
-    readAppFontFamily(),
-  );
-  // Surface (light/dark) tracks the same resolution `applyAppTheme()` uses,
-  // so the preview card mirrors what the app actually paints. For `auto`,
-  // we listen to `prefers-color-scheme` so OS flips refresh the preview
-  // without closing the modal.
-  const [surface, setSurface] = useState<"light" | "dark">(() =>
-    resolveAppSurface(readAppTheme()),
-  );
-  useEffect(() => {
-    setSurface(resolveAppSurface(theme));
-    if (theme !== "auto") return;
-    let media: MediaQueryList | null = null;
-    try {
-      media = window.matchMedia("(prefers-color-scheme: light)");
-    } catch {
-      return;
-    }
-    const onChange = () => setSurface(resolveAppSurface("auto"));
-    if (media.addEventListener) {
-      media.addEventListener("change", onChange);
-      return () => media!.removeEventListener("change", onChange);
-    }
-    media.addListener(onChange);
-    return () => media!.removeListener(onChange);
-  }, [theme]);
-  const setTheme = (next: AppTheme) => {
-    setThemeState(next);
-    writeAppTheme(next);
-    notifySameWindowStorage(STORAGE_APP_THEME, next);
-  };
-  const setLightVariant = (next: LightVariant) => {
-    setLightVariantState(next);
-    writeLightVariant(next);
-    notifySameWindowStorage(STORAGE_APP_LIGHT_VARIANT, next);
-  };
-  const setDarkVariant = (next: DarkVariant) => {
-    setDarkVariantState(next);
-    writeDarkVariant(next);
-    notifySameWindowStorage(STORAGE_APP_DARK_VARIANT, next);
-  };
-  const setFontFamily = (next: AppFontFamily) => {
-    setFontFamilyState(next);
-    writeAppFontFamily(next);
-    notifySameWindowStorage(STORAGE_APP_FONT_FAMILY, next);
-  };
-  return (
-    <>
-      <PaneHeader
-        title="Appearance"
-        subtitle="Theme, palette, and font."
-      />
-      <Row
-        label="Theme"
-        sub="Match the OS, or pin to light or dark."
-      >
-        <ThemeSegmented value={theme} onChange={setTheme} />
-      </Row>
-      <Row
-        label="Light theme"
-        sub="Picked when the OS is light or Theme = Light."
-      >
-        <StyledSelect
-          value={lightVariant}
-          options={LIGHT_VARIANT_OPTIONS.map((id) => ({
-            value: id,
-            label: LIGHT_VARIANT_LABELS[id],
-            swatchColor: LIGHT_VARIANT_ACCENTS[id],
-          }))}
-          onChange={(v) => setLightVariant(v as LightVariant)}
-        />
-      </Row>
-      <Row
-        label="Dark theme"
-        sub="Picked when the OS is dark or Theme = Dark."
-      >
-        <StyledSelect
-          value={darkVariant}
-          options={DARK_VARIANT_OPTIONS.map((id) => ({
-            value: id,
-            label: DARK_VARIANT_LABELS[id],
-            swatchColor: DARK_VARIANT_ACCENTS[id],
-          }))}
-          onChange={(v) => setDarkVariant(v as DarkVariant)}
-        />
-      </Row>
-      <Row
-        label="App font"
-        sub="UI typeface across the app. Doesn't apply to the embedded terminal — see Terminal pane."
-      >
-        <StyledSelect
-          value={fontFamily}
-          options={APP_FONT_FAMILY_OPTIONS.map((f) => ({
-            value: f,
-            label: f,
-          }))}
-          onChange={(v) => setFontFamily(v as AppFontFamily)}
-        />
-      </Row>
-      <div className="h-px w-full bg-line" />
-      <AppearancePreview
-        surface={surface}
-        lightVariant={lightVariant}
-        darkVariant={darkVariant}
-      />
-    </>
-  );
-}
-
-// Preview card. Mirrors the Pencil node `r7Su0` in `runners-design.pen`
-// — a section header + a mini chat-row card the user can compare-shop
-// with before committing the dropdown. The inner wrapper sets
-// `data-theme` to the active variant for the resolved surface, with
-// one exception: Carbon (the canonical dark variant) leaves the
-// attribute off so the unattributed `@theme` defaults paint through
-// — mirrors `applyAppTheme()`'s cascade.
-function AppearancePreview({
-  surface,
-  lightVariant,
-  darkVariant,
-}: {
-  surface: "light" | "dark";
-  lightVariant: LightVariant;
-  darkVariant: DarkVariant;
-}) {
-  const previewTheme: string | undefined =
-    surface === "light"
-      ? lightVariant
-      : darkVariant === "carbon"
-        ? undefined
-        : darkVariant;
-  return (
-    <div className="flex flex-col gap-2">
-      <div className="font-mono text-[10px] font-semibold uppercase tracking-[0.15em] text-fg-3">
-        Preview
-      </div>
-      <div
-        data-theme={previewTheme}
-        className="flex items-center gap-3 rounded-lg border border-line bg-raised p-3"
-      >
-        <div className="flex min-w-0 flex-1 flex-col gap-0.5">
-          <span className="truncate text-[13px] font-semibold text-fg">
-            @architect
-          </span>
-          <span className="truncate text-[11px] text-fg-2">
-            Plans features and splits work into tasks.
-          </span>
-        </div>
-        <button
-          type="button"
-          className="shrink-0 cursor-default rounded-md bg-accent px-3 py-1.5 text-[12px] font-medium text-accent-ink"
-        >
-          Chat
-        </button>
-      </div>
-    </div>
-  );
-}
-
-// Segmented Auto · Light · Dark control. Mirrors the Pencil node `J0lKR`
-// in `runners-design.pen` — 3 cells inside a rounded container, active
-// cell carries the raised surface + accent label, inactive cells stay
-// muted on the bg.
-function ThemeSegmented({
-  value,
-  onChange,
-}: {
-  value: AppTheme;
-  onChange: (next: AppTheme) => void;
-}) {
-  const ICONS: Record<AppTheme, typeof Monitor> = {
-    auto: Monitor,
-    light: Sun,
-    dark: Moon,
-  };
-  const LABELS: Record<AppTheme, string> = {
-    auto: "Auto",
-    light: "Light",
-    dark: "Dark",
-  };
-  return (
-    <div
-      role="radiogroup"
-      aria-label="Theme"
-      className="flex items-center gap-0.5 rounded-md border border-line bg-panel p-0.5"
-    >
-      {APP_THEME_OPTIONS.map((option) => {
-        const Icon = ICONS[option];
-        const active = option === value;
-        return (
-          <button
-            key={option}
-            type="button"
-            role="radio"
-            aria-checked={active}
-            onClick={() => onChange(option)}
-            className={`flex cursor-pointer items-center gap-1.5 rounded-[4px] px-2.5 py-[5px] text-[12px] font-medium transition-colors ${
-              active
-                ? "bg-raised text-fg"
-                : "text-fg-2 hover:text-fg"
-            }`}
-          >
-            <Icon aria-hidden className="h-3 w-3" />
-            <span>{LABELS[option]}</span>
-          </button>
-        );
-      })}
-    </div>
-  );
-}
-
-function TerminalPane() {
-  const [fontSize, setFontSizeState] = useState<number>(() =>
-    readTerminalFontSize(),
-  );
-  const [fontFamily, setFontFamilyState] = useState<TerminalFontFamily>(() =>
-    readTerminalFontFamily(),
-  );
-  const [cursorStyle, setCursorStyleState] = useState<TerminalCursorStyle>(
-    () => readTerminalCursorStyle(),
-  );
-  const [scrollback, setScrollbackState] = useState<number>(() =>
-    readTerminalScrollback(),
-  );
-  const [theme, setThemeState] = useState<TerminalTheme>(() =>
-    readTerminalTheme(),
-  );
-  const setFontSize = (next: number) => {
-    setFontSizeState(next);
-    writeTerminalFontSize(next);
-    notifySameWindowStorage(STORAGE_TERMINAL_FONT_SIZE, String(next));
-  };
-  const setFontFamily = (next: TerminalFontFamily) => {
-    setFontFamilyState(next);
-    writeTerminalFontFamily(next);
-    notifySameWindowStorage(STORAGE_TERMINAL_FONT_FAMILY, next);
-  };
-  const setCursorStyle = (next: TerminalCursorStyle) => {
-    setCursorStyleState(next);
-    writeTerminalCursorStyle(next);
-    notifySameWindowStorage(STORAGE_TERMINAL_CURSOR_STYLE, next);
-  };
-  const setScrollback = (next: number) => {
-    setScrollbackState(next);
-    writeTerminalScrollback(next);
-    notifySameWindowStorage(STORAGE_TERMINAL_SCROLLBACK, String(next));
-  };
-  const setTheme = (next: TerminalTheme) => {
-    setThemeState(next);
-    writeTerminalTheme(next);
-    notifySameWindowStorage(STORAGE_TERMINAL_THEME, next);
-  };
-  return (
-    <>
-      <PaneHeader
-        title="Terminal"
-        subtitle="xterm appearance settings for the runner terminal."
-      />
-      <Row
-        label="Theme"
-        sub="ANSI palette for the embedded terminal."
-      >
-        <StyledSelect
-          value={theme}
-          options={TERMINAL_THEME_OPTIONS.map((id) => ({
-            value: id,
-            label: TERMINAL_THEME_LABELS[id],
-            swatchColor: TERMINAL_THEME_ACCENTS[id],
-          }))}
-          onChange={(v) => setTheme(v as TerminalTheme)}
-        />
-      </Row>
-      <Row
-        label="Font family"
-        sub="Typeface used by the embedded terminal."
-      >
-        <StyledSelect
-          value={fontFamily}
-          options={TERMINAL_FONT_FAMILY_OPTIONS.map((f) => ({
-            value: f,
-            label: f,
-          }))}
-          onChange={(v) => setFontFamily(v as TerminalFontFamily)}
-        />
-      </Row>
-      <Row
-        label="Terminal font size"
-        sub="Glyph size for the embedded terminal."
-      >
-        <FontSizeStepper value={fontSize} onChange={setFontSize} />
-      </Row>
-      <Row
-        label="Cursor style"
-        sub="Block, underline, or bar — affects the prompt caret only."
-      >
-        <StyledSelect
-          value={cursorStyle}
-          options={TERMINAL_CURSOR_STYLE_OPTIONS.map((c) => ({
-            value: c,
-            label: c[0].toUpperCase() + c.slice(1),
-          }))}
-          onChange={(v) => setCursorStyle(v as TerminalCursorStyle)}
-        />
-      </Row>
-      <Row
-        label="Scrollback"
-        sub="Lines kept in history per session. Higher uses more memory."
-      >
-        <div className="flex items-center gap-2">
-          <StyledSelect
-            value={String(scrollback)}
-            options={TERMINAL_SCROLLBACK_OPTIONS.map((n) => ({
-              value: String(n),
-              label: n.toLocaleString(),
-            }))}
-            onChange={(v) => setScrollback(Number.parseInt(v, 10))}
-          />
-          <span className="text-[12px] text-fg-2">lines</span>
-        </div>
-      </Row>
-    </>
-  );
-}
-
-// Generic [−] <value> [+] stepper matching pencil nodes `KfYfw` and `CcFty`.
-// Caller renders the value cell's contents and supplies its width.
-function Stepper({
-  valueCellWidth,
-  decDisabled,
-  incDisabled,
-  onDec,
-  onInc,
-  decAriaLabel,
-  incAriaLabel,
-  children,
-}: {
-  valueCellWidth: number;
-  decDisabled?: boolean;
-  incDisabled?: boolean;
-  onDec: () => void;
-  onInc: () => void;
-  decAriaLabel: string;
-  incAriaLabel: string;
-  children: React.ReactNode;
-}) {
-  const buttonClass =
-    "flex h-[30px] w-[30px] shrink-0 cursor-pointer items-center justify-center text-fg-3 transition-colors hover:text-fg disabled:cursor-not-allowed disabled:opacity-40 disabled:hover:text-fg-3";
-  return (
-    <div className="flex h-[30px] items-center rounded-md border border-line bg-bg">
-      <button
-        type="button"
-        onClick={onDec}
-        disabled={decDisabled}
-        aria-label={decAriaLabel}
-        className={buttonClass}
-      >
-        <Minus aria-hidden className="h-3.5 w-3.5" />
-      </button>
-      <div
-        style={{ width: valueCellWidth }}
-        className="flex h-[30px] items-center justify-center border-x border-line"
-      >
-        {children}
-      </div>
-      <button
-        type="button"
-        onClick={onInc}
-        disabled={incDisabled}
-        aria-label={incAriaLabel}
-        className={buttonClass}
-      >
-        <Plus aria-hidden className="h-3.5 w-3.5" />
-      </button>
-    </div>
-  );
-}
-
-function ZoomStepper({
-  value,
-  onChange,
-}: {
-  value: number;
-  onChange: (v: number) => void;
-}) {
-  // Snap a possibly-stale persisted value to the nearest known step so the
-  // user can always move with `−`/`+`; nothing in the modal hard-blocks
-  // off-step values.
-  const idx = ZOOM_STEPS.findIndex((s) => Math.abs(s - value) < 0.001);
-  const currentIdx = idx === -1 ? ZOOM_STEPS.indexOf(1.0) : idx;
-  const pct = Math.round(ZOOM_STEPS[currentIdx] * 100);
-  return (
-    <Stepper
-      valueCellWidth={56}
-      decDisabled={currentIdx <= 0}
-      incDisabled={currentIdx >= ZOOM_STEPS.length - 1}
-      decAriaLabel="Decrease zoom"
-      incAriaLabel="Increase zoom"
-      onDec={() => onChange(ZOOM_STEPS[Math.max(0, currentIdx - 1)])}
-      onInc={() =>
-        onChange(ZOOM_STEPS[Math.min(ZOOM_STEPS.length - 1, currentIdx + 1)])
-      }
-    >
-      <span className="font-mono text-[12px] font-medium text-fg">{pct}%</span>
-    </Stepper>
-  );
-}
-
-function FontSizeStepper({
-  value,
-  onChange,
-}: {
-  value: number;
-  onChange: (v: number) => void;
-}) {
-  const clamped = Math.max(
-    TERMINAL_FONT_SIZE_MIN,
-    Math.min(TERMINAL_FONT_SIZE_MAX, value),
-  );
-  return (
-    <Stepper
-      valueCellWidth={64}
-      decDisabled={clamped <= TERMINAL_FONT_SIZE_MIN}
-      incDisabled={clamped >= TERMINAL_FONT_SIZE_MAX}
-      decAriaLabel="Decrease terminal font size"
-      incAriaLabel="Increase terminal font size"
-      onDec={() => onChange(Math.max(TERMINAL_FONT_SIZE_MIN, clamped - 1))}
-      onInc={() => onChange(Math.min(TERMINAL_FONT_SIZE_MAX, clamped + 1))}
-    >
-      <span className="flex items-center gap-[3px]">
-        <span className="font-mono text-[12px] font-medium text-fg">
-          {clamped}
-        </span>
-        <span className="font-mono text-[10px] text-fg-3">px</span>
-      </span>
-    </Stepper>
-  );
-}
-
-function McpPane() {
-  const [status, setStatus] = useState<McpIntegrationStatus | null>(null);
-  const [snippet, setSnippet] = useState<McpConfigSnippet | null>(null);
-  const [busy, setBusy] = useState<McpClientId | null>(null);
-  const [copied, setCopied] = useState<McpCopyKey | null>(null);
-  const [error, setError] = useState<string | null>(null);
-  const bindingDir = status?.socket_path
-    ? parentPath(status.socket_path)
-    : "";
-
-  const refresh = async () => {
-    try {
-      const [nextStatus, nextSnippet] = await Promise.all([
-        invoke<McpIntegrationStatus>("mcp_integration_status"),
-        invoke<McpConfigSnippet>("mcp_config_snippet"),
-      ]);
-      setStatus(nextStatus);
-      setSnippet(nextSnippet);
-      setError(null);
-    } catch (e) {
-      setError(e instanceof Error ? e.message : String(e));
-    }
-  };
-
-  useEffect(() => {
-    void refresh();
-  }, []);
-
-  const setIntegration = async (client: McpClientId, enabled: boolean) => {
-    setBusy(client);
-    setError(null);
-    try {
-      await invoke("mcp_set_integration", { client, enabled });
-      await refresh();
-    } catch (e) {
-      setError(e instanceof Error ? e.message : String(e));
-    } finally {
-      setBusy(null);
-    }
-  };
-
-  const copyMcpText = async (key: McpCopyKey, text?: string) => {
-    if (!text) return;
-    try {
-      await navigator.clipboard.writeText(text);
-      setCopied(key);
-      window.setTimeout(() => setCopied(null), 1500);
-    } catch (e) {
-      setError(e instanceof Error ? e.message : String(e));
-    }
-  };
-
-  return (
-    <>
-      <PaneHeader
-        title="MCP"
-        subtitle="Register Runner with external MCP clients."
-      />
-      <div className="rounded-lg border border-line bg-bg p-3.5">
-        <div className="mb-2 flex items-center justify-between gap-3">
-          <div className="flex items-center gap-2">
-            <Plug aria-hidden className="h-3.5 w-3.5 text-fg-2" />
-            <span className="text-[13px] font-semibold text-fg">
-              Current binding
-            </span>
-          </div>
-          <McpEnvironmentBadge environment={status?.environment} />
-        </div>
-        <BindingLine
-          label="Binding dir"
-          value={bindingDir}
-          copied={copied === "binding_dir"}
-          onCopy={() => void copyMcpText("binding_dir", bindingDir)}
-        />
-      </div>
-      <McpClientRow
-        title="Claude Code"
-        subtitle="Writes the runner entry under ~/.claude.json."
-        status={status?.claude_code ?? null}
-        busy={busy === "claude_code"}
-        onToggle={(enabled) => void setIntegration("claude_code", enabled)}
-      />
-      <McpClientRow
-        title="Codex CLI"
-        subtitle="Writes the runner table under ~/.codex/config.toml."
-        status={status?.codex ?? null}
-        busy={busy === "codex"}
-        onToggle={(enabled) => void setIntegration("codex", enabled)}
-      />
-      <div className="rounded-lg border border-line bg-bg p-3.5">
-        <div className="mb-2 flex items-center justify-between gap-3">
-          <span className="text-[13px] font-semibold text-fg">
-            Manual config
-          </span>
-          <div className="flex items-center gap-2">
-            <CopyButton
-              copied={copied === "claude_code"}
-              label="Claude"
-              onClick={() => void copyMcpText("claude_code", snippet?.claude_code)}
-            />
-            <CopyButton
-              copied={copied === "codex"}
-              label="Codex"
-              onClick={() => void copyMcpText("codex", snippet?.codex)}
-            />
-          </div>
-        </div>
-        <p className="text-[11px] leading-[1.5] text-fg-2">
-          Use these snippets for clients Runner does not update directly.
-        </p>
-      </div>
-      <div className="flex items-start gap-2.5 rounded-lg border border-line bg-panel px-3 py-2.5">
-        <ShieldAlert aria-hidden className="mt-0.5 h-3.5 w-3.5 shrink-0 text-accent" />
-        <p className="text-[11px] leading-[1.5] text-fg-2">
-          Registering replaces only the `runner` MCP entry. If the row says it
-          points to another binary, replacing it will move that client to the
-          binding shown above.
-        </p>
-      </div>
-      {error ? (
-        <div className="flex items-start justify-between gap-3 rounded-lg border border-red-500/30 bg-red-500/10 px-3.5 py-2.5">
-          <p className="min-w-0 text-[12px] text-red-300">{error}</p>
-          <button
-            type="button"
-            onClick={() => void refresh()}
-            className="shrink-0 cursor-pointer text-[12px] font-medium text-red-200 underline"
-          >
-            Retry
-          </button>
-        </div>
-      ) : null}
-    </>
-  );
-}
-
-function McpEnvironmentBadge({ environment }: { environment?: string }) {
-  const loading = !environment;
-  const development = environment?.toLowerCase().includes("dev") ?? false;
-
-  return (
-    <span
-      className={`inline-flex items-center gap-1.5 rounded px-2 py-0.5 text-[10px] font-medium ${
-        loading
-          ? "bg-raised text-fg-3"
-          : development
-            ? "bg-amber-400/10 text-amber-200"
-            : "bg-accent/10 text-accent"
-      }`}
-    >
-      <span
-        className={`h-1.5 w-1.5 rounded-full ${
-          loading
-            ? "bg-fg-3"
-            : development
-              ? "bg-amber-300"
-              : "bg-accent"
-        }`}
-      />
-      {environment ?? "Loading"}
-    </span>
-  );
-}
-
-function parentPath(path: string) {
-  const index = path.lastIndexOf("/");
-  return index > 0 ? path.slice(0, index) : path;
-}
-
-function BindingLine({
-  label,
-  value,
-  copied,
-  onCopy,
-}: {
-  label: string;
-  value: string;
-  copied: boolean;
-  onCopy: () => void;
-}) {
-  return (
-    <div className="grid grid-cols-[72px_minmax(0,1fr)] items-center gap-2 py-1">
-      <span className="text-[11px] text-fg-3">{label}</span>
-      <div className="group relative flex h-8 min-w-0 items-center gap-2 rounded bg-raised px-2">
-        <span className="min-w-0 flex-1 truncate font-mono text-[11px] text-fg-2">
-          {value || "Loading..."}
-        </span>
-        {value ? (
-          <div className="pointer-events-none absolute left-0 top-full z-50 mt-1 hidden max-w-[560px] rounded border border-line bg-raised px-2 py-1.5 font-mono text-[10px] leading-[1.45] text-fg shadow-[0_8px_24px_rgba(0,0,0,0.5)] group-hover:block group-focus-within:block">
-            {value}
-          </div>
-        ) : null}
-        <button
-          type="button"
-          aria-label={`Copy ${label}`}
-          disabled={!value}
-          onClick={onCopy}
-          className="flex h-5 w-5 shrink-0 cursor-pointer items-center justify-center rounded text-fg-3 transition-colors hover:bg-line/60 hover:text-fg disabled:cursor-not-allowed disabled:opacity-40"
-        >
-          {copied ? (
-            <Check aria-hidden className="h-3 w-3" />
-          ) : (
-            <Copy aria-hidden className="h-3 w-3" />
-          )}
-        </button>
-      </div>
-    </div>
-  );
-}
-
-function McpClientRow({
-  title,
-  subtitle,
-  status,
-  busy,
-  onToggle,
-}: {
-  title: string;
-  subtitle: string;
-  status: McpClientStatus | null;
-  busy: boolean;
-  onToggle: (enabled: boolean) => void;
-}) {
-  const loading = status == null;
-  const hasError = Boolean(status?.error);
-  const active = status?.matches_current ?? false;
-  const stateLabel = loading
-    ? "Checking"
-    : busy
-      ? "Updating"
-    : hasError
-      ? "Config error"
-      : !status.registered
-        ? "Not registered"
-        : status.matches_current
-          ? ""
-          : "Registered to another Runner";
-  const stateClass = loading
-    ? "text-fg-3"
-    : hasError
-      ? "text-red-300"
-      : status?.matches_current
-        ? "text-accent"
-      : status?.registered
-        ? "text-amber-300"
-        : "text-fg-3";
-  const configured = status?.registered
-    ? `${status.command ?? "(missing command)"} ${JSON.stringify(status.args)}`
-    : "";
-  const showDetail = hasError || Boolean(status?.registered && !status.matches_current);
-
-  return (
-    <div className="rounded-lg border border-line bg-bg p-3.5">
-      <div className="flex items-start justify-between gap-4">
-        <div className="min-w-0">
-          <div className="flex items-center gap-2">
-            <span className="text-[13px] font-semibold text-fg">{title}</span>
-            {stateLabel ? (
-              <span className={`text-[11px] ${stateClass}`}>{stateLabel}</span>
-            ) : null}
-          </div>
-          <p className="mt-0.5 text-[11px] text-fg-2">{subtitle}</p>
-        </div>
-        <Toggle
-          on={active}
-          onChange={onToggle}
-          disabled={loading || busy || hasError}
-        />
-      </div>
-      {showDetail ? (
-        <div className="mt-2 min-w-0 rounded bg-raised px-2 py-1.5">
-          <McpStatusLine
-            label={status?.error ? "Error" : "Configured command"}
-            value={status?.error ?? configured}
-          />
-        </div>
-      ) : null}
-    </div>
-  );
-}
-
-function McpStatusLine({ label, value }: { label: string; value: string }) {
-  return (
-    <div className="grid grid-cols-[112px_minmax(0,1fr)] gap-2">
-      <span className="text-[10px] text-fg-3">{label}</span>
-      <span className="break-all font-mono text-[10px] leading-[1.45] text-fg-3">
-        {value}
-      </span>
-    </div>
-  );
-}
-
-function CopyButton({
-  copied,
-  label,
-  onClick,
-}: {
-  copied: boolean;
-  label: string;
-  onClick: () => void;
-}) {
-  return (
-    <button
-      type="button"
-      onClick={onClick}
-      className="flex cursor-pointer items-center gap-1.5 rounded-md border border-line bg-panel px-2.5 py-1 text-[12px] font-medium text-fg-2 transition-colors hover:border-line-strong hover:text-fg"
-    >
-      {copied ? (
-        <Check aria-hidden className="h-3 w-3" />
-      ) : (
-        <Copy aria-hidden className="h-3 w-3" />
-      )}
-      {copied ? "Copied" : label}
-    </button>
-  );
-}
-
-function DiagnosticsPane() {
-  // Mirrors the Help → "Reveal logs in Finder" menu item. Both routes
-  // invoke `runner_logs_reveal`, which resolves
-  // `~/Library/Logs/com.wycstudios.runner/` and hands it to the
-  // opener plugin. Dedicated pane (not buried in About) so #10/#13/#14
-  // diagnostic affordances have a home to grow into.
-  const [busy, setBusy] = useState(false);
-  const reveal = async () => {
-    if (busy) return;
-    setBusy(true);
-    try {
-      await invoke("runner_logs_reveal");
-    } catch (e) {
-      // Best-effort: the only failure path is opener refusing the
-      // path or the log dir not yet existing. The backend creates
-      // the dir on demand, so a swallow here keeps the modal quiet —
-      // a follow-up could surface a toast.
-      console.error("reveal logs failed", e);
-    } finally {
-      setBusy(false);
-    }
-  };
-  return (
-    <>
-      <PaneHeader
-        title="Diagnostics"
-        subtitle="Logs and troubleshooting tools."
-      />
-      <Row
-        label="Application logs"
-        sub="Open the folder containing runner.log so you can attach it to a bug report."
-      >
-        <button
-          type="button"
-          onClick={() => void reveal()}
-          disabled={busy}
-          className="flex shrink-0 cursor-pointer items-center gap-1.5 whitespace-nowrap rounded-md border border-line bg-panel px-3 py-1.5 text-[12px] font-medium text-fg-2 transition-colors hover:border-line-strong hover:text-fg disabled:cursor-not-allowed disabled:opacity-60"
-        >
-          <FolderOpen aria-hidden className="h-3 w-3" />
-          Reveal logs in Finder
-        </button>
-      </Row>
-    </>
-  );
-}
-
-function UpdatesPane() {
-  const [autoInstall, setAutoInstall] = useStoredBool(
-    STORAGE_AUTO_INSTALL_UPDATES,
-    true,
-  );
-  const [version, setVersion] = useState<string>("");
-  const {
-    status,
-    update,
-    progress,
-    error,
-    checkForUpdate,
-    downloadAndInstall,
-    restart,
-  } = useUpdate();
-  useEffect(() => {
-    void getVersion()
-      .then((v) => setVersion(v))
-      .catch(() => setVersion(""));
-  }, []);
-  // Status-driven copy + action — mirrors the design's three active
-  // panes (`pYv9W` available, `u4odWB` downloading, `KVWlJ` ready).
-  // The same row swaps into each state so users always see ONE
-  // current action.
-  const sub = (() => {
-    switch (status) {
-      case "checking":
-        return "Checking for updates…";
-      case "available":
-        return update?.version
-          ? `v${update.version} is available.`
-          : "An update is available.";
-      case "downloading":
-        return `Downloading update… ${progress}%`;
-      case "ready":
-        return "Update ready — restart to apply.";
-      case "error":
-        return error ? `Couldn't check: ${error}` : "Couldn't check for updates.";
-      default:
-        return "You're up to date.";
-    }
-  })();
-  const subTone =
-    status === "available" || status === "ready"
-      ? "text-accent"
-      : "text-fg-2";
-  return (
-    <>
-      <PaneHeader
-        title="Updates"
-        subtitle="Stay current with the latest version."
-      />
-      {/* Version card — same shell, different right-hand action per state. */}
-      <div className="flex items-center justify-between gap-4 rounded-lg border border-line bg-bg p-3.5">
-        <div className="flex min-w-0 flex-col gap-0.5">
-          <div className="flex items-center gap-2">
-            <span className="text-[13px] font-semibold text-fg">Runner</span>
-            <span className="rounded bg-raised px-1.5 py-0.5 font-mono text-[10px] text-fg-2">
-              v{version || "0.0.0"}
-            </span>
-          </div>
-          <span className={`truncate text-[11px] ${subTone}`}>{sub}</span>
-        </div>
-        <UpdatesAction
-          status={status}
-          onCheck={() => void checkForUpdate()}
-          onDownload={() => void downloadAndInstall()}
-          onRestart={() => void restart()}
-        />
-      </div>
-      {status === "downloading" ? (
-        <div className="-mt-2 h-[3px] w-full overflow-hidden rounded-full bg-raised">
-          <div
-            className="h-full rounded-full bg-accent transition-[width] duration-200"
-            style={{ width: `${progress}%` }}
-          />
-        </div>
-      ) : null}
-      <Row
-        label="Install updates automatically"
-        sub="Download and apply updates in the background. Restart needed to finish."
-      >
-        <Toggle on={autoInstall} onChange={setAutoInstall} />
-      </Row>
-    </>
-  );
-}
-
-function UpdatesAction({
-  status,
-  onCheck,
-  onDownload,
-  onRestart,
-}: {
-  status: ReturnType<typeof useUpdate>["status"];
-  onCheck: () => void;
-  onDownload: () => void;
-  onRestart: () => void;
-}) {
-  // No button slot during download — the inline progress bar is the
-  // affordance. Cancel isn't supported by the plugin (no abort
-  // handle), so we don't pretend to offer one.
-  if (status === "checking") {
-    return (
-      <div className="flex shrink-0 items-center gap-1.5 whitespace-nowrap rounded-md border border-line bg-panel px-3 py-1.5 text-[12px] font-medium text-fg-2">
-        <Loader2 aria-hidden className="h-3 w-3 animate-spin" />
-        Checking…
-      </div>
-    );
-  }
-  if (status === "available") {
-    return (
-      <button
-        type="button"
-        onClick={onDownload}
-        className="flex shrink-0 cursor-pointer items-center gap-1.5 whitespace-nowrap rounded-md bg-accent px-3 py-1.5 text-[12px] font-semibold text-bg transition-colors hover:bg-accent/90"
-      >
-        <Download aria-hidden className="h-3 w-3" />
-        Download &amp; install
-      </button>
-    );
-  }
-  if (status === "downloading") {
-    return (
-      <div className="flex shrink-0 items-center gap-1.5 whitespace-nowrap rounded-md border border-line bg-panel px-3 py-1.5 text-[12px] font-medium text-fg-2">
-        <Loader2 aria-hidden className="h-3 w-3 animate-spin" />
-        Installing
-      </div>
-    );
-  }
-  if (status === "ready") {
-    return (
-      <button
-        type="button"
-        onClick={onRestart}
-        className="flex shrink-0 cursor-pointer items-center gap-1.5 whitespace-nowrap rounded-md bg-accent px-3 py-1.5 text-[12px] font-semibold text-bg transition-colors hover:bg-accent/90"
-      >
-        <RotateCcw aria-hidden className="h-3 w-3" />
-        Restart to update
-      </button>
-    );
-  }
-  // idle / error → manual re-check
-  return (
-    <button
-      type="button"
-      onClick={onCheck}
-      className="flex shrink-0 cursor-pointer items-center gap-1.5 whitespace-nowrap rounded-md border border-line bg-panel px-3 py-1.5 text-[12px] font-medium text-fg-2 transition-colors hover:border-line-strong hover:text-fg"
-    >
-      <RefreshCw aria-hidden className="h-3 w-3" />
-      Check now
-    </button>
-  );
-}
-
-function AboutPane() {
-  const [version, setVersion] = useState<string>("");
-  // Platform/arch label derived from navigator.userAgent — Tauri 2's
-  // dedicated `@tauri-apps/plugin-os` isn't installed yet, and the
-  // pill is informational only. Falls back to "desktop" if parsing
-  // fails.
-  const platformLabel = (() => {
-    if (typeof navigator === "undefined") return "";
-    const ua = navigator.userAgent.toLowerCase();
-    const arch = ua.includes("arm64") || ua.includes("aarch64")
-      ? "arm64"
-      : ua.includes("x86_64") || ua.includes("x64")
-        ? "x86_64"
-        : "";
-    const os = ua.includes("mac")
-      ? "darwin"
-      : ua.includes("win")
-        ? "windows"
-        : ua.includes("linux")
-          ? "linux"
-          : "";
-    if (!os) return "";
-    return arch ? `${os} · ${arch}` : os;
-  })();
-  useEffect(() => {
-    void getVersion()
-      .then((v) => setVersion(v))
-      .catch(() => setVersion(""));
-  }, []);
-  const openLink = (url: string) => {
-    void openUrl(url).catch(() => {
-      // Fallback: window.open works in dev (browser preview) when
-      // the Tauri opener plugin isn't available.
-      window.open(url, "_blank");
-    });
-  };
-  return (
-    <>
-      {/* Identity block — replaces PaneHeader because the About pane
-          centers the brand block instead of using a left-aligned
-          h2 + subtitle. */}
-      <div className="flex flex-col items-center gap-3.5 pb-2 pt-1">
-        {/* About-page icon mirrors the bundled `.icns` shipped with
-            the macOS / Windows installer instead of a stylized chevron
-            glyph, so the in-app identity matches what users see in
-            the Dock / file-explorer. Imported from `src/assets/` so
-            Vite emits a hashed URL into the bundle. */}
-        <img
-          src={appIcon}
-          alt="Runner icon"
-          width={56}
-          height={56}
-          className="h-14 w-14 rounded-2xl"
-        />
-        <div className="flex flex-col items-center gap-1.5">
-          <span className="text-[20px] font-bold text-fg">Runner</span>
-          <span className="text-[12px] text-fg-2">
-            Desktop editor for crews of local CLI coding agents.
-          </span>
-        </div>
-        <div className="flex items-center gap-2">
-          <span className="rounded bg-raised px-2 py-0.5 font-mono text-[11px] text-fg-2">
-            v{version || "0.0.0"}
-          </span>
-          {platformLabel ? (
-            <span className="rounded bg-raised px-2 py-0.5 font-mono text-[11px] text-fg-2">
-              {platformLabel}
-            </span>
-          ) : null}
-        </div>
-      </div>
-      <div className="h-px w-full bg-line" />
-      <div className="flex flex-col gap-px">
-        <LinkRow
-          icon={<GithubGlyph />}
-          label="GitHub"
-          onClick={() => openLink("https://github.com/yicheng47/runner")}
-          external
-        />
-        <LinkRow
-          icon={<BookText aria-hidden className="h-3.5 w-3.5 text-fg-2" />}
-          label="Documentation"
-          onClick={() => openLink("https://github.com/yicheng47/runner#readme")}
-          external
-        />
-        <LinkRow
-          icon={<Scale aria-hidden className="h-3.5 w-3.5 text-fg-2" />}
-          label="License"
-          trailing={<span className="text-[12px] text-fg-3">MIT</span>}
-        />
-      </div>
-      <div className="flex-1" />
-      <div className="flex items-center justify-center text-[11px] text-fg-3">
-        © 2026 wyc studios
-      </div>
-    </>
-  );
-}
-
-function Row({
-  label,
-  sub,
-  children,
-}: {
-  label: string;
-  sub?: string;
-  children: React.ReactNode;
-}) {
-  return (
-    <div className="flex items-center justify-between gap-6">
-      <div className="flex min-w-0 flex-col gap-0.5">
-        <span className="text-[13px] font-medium text-fg">{label}</span>
-        {sub ? <span className="text-[11px] text-fg-2">{sub}</span> : null}
-      </div>
-      <div className="shrink-0">{children}</div>
-    </div>
-  );
-}
-
-function Toggle({
-  on,
-  onChange,
-  disabled,
-}: {
-  on: boolean;
-  onChange: (v: boolean) => void;
-  disabled?: boolean;
-}) {
-  return (
-    <button
-      type="button"
-      role="switch"
-      aria-checked={on}
-      disabled={disabled}
-      onClick={() => {
-        if (disabled) return;
-        onChange(!on);
-      }}
-      className={`flex h-[18px] w-8 cursor-pointer items-center rounded-full p-0.5 transition-colors disabled:cursor-not-allowed disabled:opacity-50 ${
-        on ? "justify-end bg-accent/15" : "justify-start bg-raised"
-      }`}
-    >
-      <span
-        className={`block h-3.5 w-3.5 rounded-full ${
-          on ? "bg-accent" : "bg-fg-3"
-        }`}
-      />
-    </button>
-  );
-}
-
-// Typed input + Browse… button — matches the working-directory
-// control in StartMissionModal / CreateRunnerModal so the three
-// surfaces feel like one family. Typing an empty string flows
-// through `writeDefaultWorkingDir` (which removes the key); the
-// Browse button overlays Tauri's native directory dialog.
-function LinkRow({
-  icon,
-  label,
-  onClick,
-  external,
-  trailing,
-}: {
-  icon: React.ReactNode;
-  label: string;
-  onClick?: () => void;
-  external?: boolean;
-  trailing?: React.ReactNode;
-}) {
-  const interactive = !!onClick;
-  return (
-    <button
-      type="button"
-      onClick={onClick}
-      disabled={!interactive}
-      className={`flex w-full items-center justify-between rounded-md px-3 py-2.5 text-left ${
-        interactive ? "cursor-pointer hover:bg-raised" : "cursor-default"
-      }`}
-    >
-      <span className="flex items-center gap-2.5">
-        <span className="flex h-3.5 w-3.5 items-center justify-center text-fg-2">
-          {icon}
-        </span>
-        <span className="text-[13px] text-fg">{label}</span>
-      </span>
-      {trailing ?? (
-        external ? (
-          <ExternalLink aria-hidden className="h-3 w-3 text-fg-3" />
-        ) : null
-      )}
-    </button>
-  );
-}
-
-function GithubGlyph() {
-  return (
-    <svg
-      width="14"
-      height="14"
-      viewBox="0 0 24 24"
-      fill="currentColor"
-      className="text-fg-2"
-      aria-hidden
-    >
-      <path d="M12 .5C5.6.5.5 5.7.5 12.1c0 5.1 3.3 9.4 7.9 10.9.6.1.8-.3.8-.6v-2.1c-3.2.7-3.9-1.5-3.9-1.5-.5-1.3-1.3-1.7-1.3-1.7-1-.7.1-.7.1-.7 1.1.1 1.7 1.2 1.7 1.2 1 1.7 2.7 1.2 3.4.9.1-.7.4-1.2.7-1.5-2.6-.3-5.3-1.3-5.3-5.7 0-1.3.4-2.3 1.2-3.1-.1-.3-.5-1.5.1-3.2 0 0 1-.3 3.3 1.2.9-.3 1.9-.4 2.9-.4s2 .1 2.9.4c2.3-1.5 3.3-1.2 3.3-1.2.7 1.7.2 2.9.1 3.2.8.8 1.2 1.9 1.2 3.1 0 4.4-2.7 5.4-5.3 5.7.4.4.8 1.1.8 2.2v3.3c0 .3.2.7.8.6 4.6-1.5 7.9-5.8 7.9-10.9C23.5 5.7 18.4.5 12 .5z" />
-    </svg>
-  );
-}
